+++ config/webpack/webpack.config.js
@@ -195,7 +195,7 @@ const defaultWebpackConfig = (env = {}, argv = {}) => {
   const resolveModules = ['node_modules'];
 
   if (!disableAggregateTranslations) {
-    aggregateTranslations(Object.assign({}, { baseDir: rootPath }, env.aggregateOptions));
+    aggregateTranslations({ baseDir: rootPath, ...env.aggregateOptions });
     resolveModules.unshift(path.resolve(rootPath, 'aggregated-translations'));
   }