+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Unreleased
 ----------
 ### Added
 * Added chunk filename support
-* Added reusable webpack entry points object
+* Added reusable webpack entry points
 
 5.4.0 - (July 31, 2019)
 ----------