+++ CHANGELOG.md
@@ -6,6 +6,9 @@ Unreleased
 ### Added
 * Added chunk filename support
 
+### Changed
+* Updated Object.assign syntax to Object spread syntax
+
 5.4.0 - (July 31, 2019)
 ----------
 ### Added